.site/requirements.txt
# Extra indexes currently necessary, because some (older) Addons have
# dependencies to packages not yet on pypi.
--extra-index-url https://devpi.divio.ch/django-cms/pre-release/+simple/
--extra-index-url https://devpi.divio.ch/divio/all/+simple/
--extra-index-url https://divio:wlkdjeorijlerf@pkg.divio.ch/simple/

-r generated_requirements.txt

Django==1.6.11  # rq.filter: <1.7
BeautifulSoup==3.2.1
Django-Select2==4.3.1
GitPython==0.3.2.RC1
Pillow==2.8.1
PyYAML==3.11
South==1.0.2
Unidecode==0.04.17
YURL==0.13
aldryn-boilerplates==0.6  # rq.filter: !=0.7
aldryn-client==0.9.9
aldryn-common==0.1.3
aldryn-sites==0.5
aldryn-snake==0.0.2
argparse==1.3.0
certifi==2015.04.28
cffi==0.9.2
chardet==2.3.0
cmsplugin-filer==0.10.1
cryptography==0.9
dj-database-url==0.3.0
dj-email-url==0.0.4
dj-redis-url==0.1.4
dj-static==0.0.6
django-appconf==1.0.1
django-cache-url==0.8.0
django-classy-tags==0.6.1
django-cms==3.1   # rq.filter: <3.2
django-compressor==1.5
django-dotenv==1.3.0
django-filer==0.9.9
django-getenv==1.3.1
django-health-check==1.0.2
django-hvad==1.2.1
django-mptt==0.6.1  # rq.filter: <0.7
django-parler==1.4
django-polymorphic==0.7.1
django-redis-cache==0.13.1
django-reversion==1.8.5  # rq.filter: !=1.8.6
django-robots==1.1
django-sekizai==0.8.1
django-simple-captcha==0.4.5
django-simple-sso==0.9.3
django-sortedm2m==0.9.5
django-treebeard==3.0
djangocms-admin-style==0.2.5
djangocms-googlemap==0.2
djangocms-link==1.6.1
djangocms-snippet==1.5
djangocms-text-ckeditor==2.4.3
djeese-fs==0.1.20
docopt==0.6.2
easy-thumbnails==2.2
enum34==1.0.4
gevent==1.0.1
gitdb==0.6.4
greenlet==0.4.7
gunicorn==18.0  # rq.filter: <19.0
hiredis==0.2.0
html5lib==0.99999
idna==2.0
ipaddress==1.0.7
itsdangerous==0.24
lxml==3.4.4
ndg-httpsclient==0.4.0
opbeat==2.1
psycogreen==1.0
psycopg2==2.6
pyOpenSSL==0.15.1
pyasn1==0.1.7
pycparser==2.13
raven==5.3.1
redis==2.10.3
requests==2.7.0
six==1.9.0
smmap==0.9.0
static3==0.6.0
webservices==0.5
wsgiref==0.1.2
